+++ src/index.ts
@@ -1,6 +1,7 @@
 import "./index.css";
 
 export * from "./components/AerConfigProvider";
+export * from "./components/Avatar";
 export * from "./components/Badge";
 export * from "./components/Button";
 export * from "./components/Card";